+++ app/src/main/java/com/mbj/doeat/ui/screen/home/detail/detail_writer/PartyDetailWriterScreen.kt
@@ -22,7 +22,7 @@ import com.mbj.doeat.data.remote.model.Party
 import com.mbj.doeat.ui.component.button.BackButton
 import com.mbj.doeat.ui.component.loading.LoadingView
 import com.mbj.doeat.ui.component.button.LongRectangleButtonWithParams
-import com.mbj.doeat.ui.component.PartyDetailContent
+import com.mbj.doeat.ui.component.party.PartyDetailContent
 import com.mbj.doeat.ui.component.ReusableWebView
 import com.mbj.doeat.ui.component.dialog.YesNoDialog
 import com.mbj.doeat.ui.screen.home.detail.detail_writer.viewmodel.PartyDetailWriterViewModel
